move default theme consts in a separate file
